refactor: extract createTestHarness + buildConfigEntry from e2e harnesses (#74)
/simplify 다중 에이전트 review가 잡은 추가 중복:

1. test runner scaffolding 4-way 중복:
   const results / let targetId / function log / async function run /
   async function test (skip 분기 포함). 4 harness 모두 동일하거나 거의
   동일. createTestHarness({getTargetId, skipSupported}) factory로
   추출. getTargetId는 getter로 받아 module-load 시점의 undefined
   targetId 문제를 회피 — findPageTarget() 결과가 도착한 뒤 closure가
   resolve. skipSupported는 run-e2e.mjs의 formula skip 패턴 한정.

2. ConfigEntry literal 19-field defaults가 3 사이트 (run-e2e /
   run-seatable-e2e / run-seatable-settings)에 거의 동일하게 중복.
   buildConfigEntry(overrides) factory가 src/types/config.types.ts의
   DEFAULT_CONFIG_ENTRY를 미러링해 plain object 반환. JSON.stringify
   해서 plugin runtime eval block에 inline. 호출 사이트는 차이만 명시
   (id / name / credentialId / tableId / folderPath + per-suite knobs).

부가 정리:
- evalInObsidian import는 obsidian-helpers.mjs로 이동 (createTestHarness
  내부에서 사용). 4 harness는 더 이상 evalInObsidian을 직접 import 안 함.
- module-level JSDoc 갱신: "String factories" + "Node-side ESM utilities"
  두 종류로 분류. make/build naming convention의 차이를 명시.

Cumulative LOC vs master:
- run-e2e.mjs:                  1230 → 1121 (-109)
- run-seatable-e2e.mjs:          759 →  657 (-102)
- run-settings-e2e.mjs:          999 →  905 (-94)
- run-seatable-settings-e2e.mjs: 413 →  308 (-105)
- obsidian-helpers.mjs (new):     —  →  320 (+320)
- 순 변화: 3401 → 3311 (-90 LOC)

검증: node --check 5 파일 / vitest 519/519 / e2e cloud 별도.
